requirements.txt
FastAPI
requests
beautifulsoup4
langchain
python-dotenv
typing
pydantic
uvicorn
langchain-openai
jinja2